Move composition wrappers next to parser
